// front_gestionOffre/src/app/core/models/candidature.model.ts
import { JobOffer } from './job.model';

export interface Candidature {
  id?: number;
  jobOffer?: JobOffer;
  candidatId?: number;
  lettreMotivation?: string;
  cvUrl?: string;
  statut?: 'EN_ATTENTE' | 'EN_COURS_EXAMEN' | 'ENTRETIEN_PLANIFIE' | 'ENTRETIEN_REALISE' | 'ACCEPTEE' | 'REFUSEE';
  datePostulation?: string;
  commentaireRecruteur?: string;
}

export interface Entretien {
  id?: number;
  candidature?: Candidature;
  statut?: 'PLANIFIE' | 'TERMINE' | 'ANNULE';
  dateHeure?: string;
  lienVisio?: string;
  dureeMinutes?: string;
  notes?: string;
}